requirements.txt
anyio==3.6.2
bleach==6.0.0
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
Flask==2.2.2
Flask-Session==0.4.0
gitdb==4.0.10
GitPython==3.1.30
gotrue==1.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
MarkupSafe==2.1.2
more-itertools==9.0.0
newsapi-python==0.2.6
packaging==23.0
pkginfo==1.9.6
postgrest-py==0.10.3
pydantic==1.10.4
Pygments==2.14.0
python-dateutil==2.8.2
python-gitlab==3.13.0
python-semantic-release==7.33.0
pywin32-ctypes==0.2.0
readme-renderer==37.3
realtime==1.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
storage3==0.5.0
supabase==1.0.0
supafunc==0.2.2
tomlkit==0.11.6
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.4.0
urllib3==1.26.14
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.12.1
